复用 knowledge 函数
